Split data into chunks to sort indevidually if possible.
Large regions were going super-slow in sort due to blowing out memory and
swapping. By splitting up the data set into chuncks we are more likely to
be able to fit the chunks into memory.

As well, if a chunk exceeds the size that we can likely do without swapping,
sorting can be skipped on that chunk.